Add complete style to the main page
Changed font family to Arial (sans-serif), for a
better readability. Wanted to add a playful character
to the page, therefore added colors. Removed margins
between elements for the complete look. Made title
heading stand out by the color choice and additional
padding. Tweaked the look of the links in <ul>,
removed dots, centered links, made them more natural
and playful by adding different colors to the
background, so as to make gradient effect
and changed the look of underline decoration.
When on hover, underline color pops out, for
intuitiveness.